Refactor `suggestion-table` into its own component #17
